update structure of how to generate routes for bllog dir
