--- docs/app/changelog/overview/page.tsx
@@ -1,348 +1,0 @@
-import { DateFormat, Headline, Text } from '@/ui';
-import { allBlogs } from 'contentlayer/generated';
-import Link from 'next/link';
-import { TeaserList } from '@/ui/TeaserCard';
-
-interface Post {
-  title: string;
-  date: Date;
-  slug: string;
-  introduction: RegExpMatchArray | null;
-}
-
-export default function Post() {
-  // matches everything till the second line break
-  const regex = /[\s\S]*?\n[\s\S]*?\n/;
-
-  const posts: Post[] = allBlogs.map(post => ({
-    title: post.title,
-    date: new Date(post.date),
-    slug: post.slug,
-    introduction: post.body.raw.match(regex),
-  }));
-
-  const sortedPosts = posts.sort((a, b) => b.date.getTime() - a.date.getTime());
-  const latestPost = sortedPosts[0];
-
-  return (
-    <article className="grid grid-cols-1 gap-x-24 gap-y-14 min-[1400px]:grid-cols-[minmax(min-content,70ch)_1fr]">
-      <div className="col-span-full">
-        <Headline level={1}>Overview</Headline>
-        <div className="text-secondary-400 pt-1">
-          Find an overview over the last Marigold release.
-        </div>
-      </div>
-      <div className="prose max-w-[70ch]">
-        {latestPost && (
-          <div key={latestPost.title} className="pb-8">
-            <Headline level={2}>
-              <Link href={`/${latestPost.slug}`}>
-                {latestPost.title} -{' '}
-                <DateFormat value={latestPost.date} dateStyle="medium" />
-              </Link>
-            </Headline>
-            <Text>{latestPost.introduction}</Text>
-          </div>
-        )}
-        <TeaserList
-          items={[
-            {
-              title: '@marigold/components',
-              href: '/packages/components/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/icons',
-              href: '/packages/icons/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/system',
-              href: '/packages/system/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/theme-preset',
-              href: '/packages/theme-preset/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/types',
-              href: '/packages/types/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/theme-b2b',
-              href: '/themes/theme-b2b/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/theme-core',
-              href: '/themes/theme-core/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/theme-docs',
-              href: '/themes/theme-docs/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/docs',
-              href: '/docs/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/eslint-config',
-              href: '/config/eslint/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/jest-config',
-              href: '/config/jest/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/prettier-config',
-              href: '/config/prettier/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/storybook-config',
-              href: '/config/storybook/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-            {
-              title: '@marigold/tsconfig-config',
-              href: '/config/tsconfig/whats-new',
-              caption: 'View the changelog of the latest release.',
-              icon: (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth={1.5}
-                  stroke="currentColor"
-                  className="size-6"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="m21 7.5-9-5.25L3 7.5m18 0-9 5.25m9-5.25v9l-9 5.25M3 7.5l9 5.25M3 7.5v9l9 5.25m0-9v9"
-                  />
-                </svg>
-              ),
-            },
-          ]}
-        />
-      </div>
-    </article>
-  );
-}
