+++ docs/content/changelog/blog-overview.mdx
@@ -1,6 +1,8 @@
 ---
 title: All Release Blogs
-caption: Find all release Blogs
+caption: Find all release blogs.
+order: 2
+badge: new
 ---
 
-/_needed for the navigation structure_/
+<PostList />
