+++ docs/content/changelog/config/jest/whats-new.mdx
@@ -1,41 +1,42 @@
 ---
 title: '@marigold/jest-config'
 caption: 'Have a look on the latest changes regarding @marigold/jest-config'
+toc: false
 ---
 
 _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blob/main/config/jest/CHANGELOG.md)_
 
-## 1.2.2 (Released on 6/28/2024)
+## 1.2.2 (Released on <DateFormat value={new Date("Fri Jun 28 2024 13:44:24 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#3967](https://github.com/marigold-ui/marigold/pull/3967) [`0773aa8`](https://github.com/marigold-ui/marigold/commit/0773aa8cd6ee71faf4f0d04f80f33cbe7fc56202) Thanks [@sebald](https://github.com/sebald)! - refa: Update TypeScript and adjust `<NumericFormat>` props
 
-## 1.2.1 (Released on 4/30/2024)
+## 1.2.1 (Released on <DateFormat value={new Date("Tue Apr 30 2024 11:05:14 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#3880](https://github.com/marigold-ui/marigold/pull/3880) [`4a59427`](https://github.com/marigold-ui/marigold/commit/4a59427fc1d4c14b1971c07778de1977cdde5dda) Thanks [@sebald](https://github.com/sebald)! - chore: cleanup peerdeeps
 
-## 1.2.0 (Released on 8/16/2023)
+## 1.2.0 (Released on <DateFormat value={new Date("Wed Aug 16 2023 13:56:21 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Minor Changes
 
 - [#3278](https://github.com/marigold-ui/marigold/pull/3278) [`74c49bfee`](https://github.com/marigold-ui/marigold/commit/74c49bfeed05cf2e20cf2c93ff59654002e4e747) Thanks [@sebald](https://github.com/sebald)! - fix: support ESM only modules better
 
-## 1.1.2 (Released on 5/16/2023)
+## 1.1.2 (Released on <DateFormat value={new Date("Tue May 16 2023 10:58:08 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2934](https://github.com/marigold-ui/marigold/pull/2934) [`41eccbce6`](https://github.com/marigold-ui/marigold/commit/41eccbce6e78e33e3c769adb5dd4b52a7ad43c92) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update testing
 
-## 1.1.1 (Released on 3/30/2023)
+## 1.1.1 (Released on <DateFormat value={new Date("Thu Mar 30 2023 09:15:15 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2887](https://github.com/marigold-ui/marigold/pull/2887) [`539544841`](https://github.com/marigold-ui/marigold/commit/53954484120a34460ddb5bb6f3b3bc7106f2ce63) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update testing
 
-## 1.1.0 (Released on 3/2/2023)
+## 1.1.0 (Released on <DateFormat value={new Date("Thu Mar 02 2023 09:07:56 GMT+0100 (Central European Standard Time)")} dateStyle="medium" />)
 
 ### Minor Changes
 
@@ -43,13 +44,13 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
 
 - [#2801](https://github.com/marigold-ui/marigold/pull/2801) [`9e682b73d`](https://github.com/marigold-ui/marigold/commit/9e682b73dafd03ca59df1f4aba71cc284a7df296) Thanks [@sebald](https://github.com/sebald)! - refa: remove ts-jest dependency
 
-## 1.0.10 (Released on 11/24/2022)
+## 1.0.10 (Released on <DateFormat value={new Date("Thu Nov 24 2022 13:48:37 GMT+0100 (Central European Standard Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2627](https://github.com/marigold-ui/marigold/pull/2627) [`d250fc00`](https://github.com/marigold-ui/marigold/commit/d250fc0041f2beca498107d3e60e2d50e9ffb293) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update babel monorepo
 
-## 1.0.9 (Released on 11/7/2022)
+## 1.0.9 (Released on <DateFormat value={new Date("Mon Nov 07 2022 13:20:21 GMT+0100 (Central European Standard Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -57,7 +58,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
 
 - [#2599](https://github.com/marigold-ui/marigold/pull/2599) [`940bc912`](https://github.com/marigold-ui/marigold/commit/940bc912d30b311e275174bc6ada34a996e8c767) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update testing
 
-## 1.0.8 (Released on 10/11/2022)
+## 1.0.8 (Released on <DateFormat value={new Date("Tue Oct 11 2022 10:29:43 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -65,19 +66,19 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
 
 - [#2545](https://github.com/marigold-ui/marigold/pull/2545) [`e65171c6`](https://github.com/marigold-ui/marigold/commit/e65171c6b30f0091491a7e0394e5ddafc0d72bf4) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update dependency @babel/core to v7.19.3
 
-## 1.0.7 (Released on 9/30/2022)
+## 1.0.7 (Released on <DateFormat value={new Date("Fri Sep 30 2022 15:12:18 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2544](https://github.com/marigold-ui/marigold/pull/2544) [`db4cb323`](https://github.com/marigold-ui/marigold/commit/db4cb323ed6b58fac1ba424ca50349bdb981bc6b) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update dependency typescript to v4.8.4
 
-## 1.0.6 (Released on 9/21/2022)
+## 1.0.6 (Released on <DateFormat value={new Date("Wed Sep 21 2022 10:23:49 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2446](https://github.com/marigold-ui/marigold/pull/2446) [`bc13abed`](https://github.com/marigold-ui/marigold/commit/bc13abedb3abc56cf6e1cc9ea8725b99c9eb9468) Thanks [@sebald](https://github.com/sebald)! - feat: losen peer deps for usage with npm
 
-## 1.0.5 (Released on 9/13/2022)
+## 1.0.5 (Released on <DateFormat value={new Date("Tue Sep 13 2022 09:42:53 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -85,7 +86,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
 
 - [#2437](https://github.com/marigold-ui/marigold/pull/2437) [`23a78264`](https://github.com/marigold-ui/marigold/commit/23a78264cf713ebf439c264a45ff946fd58472de) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update testing (major)
 
-## 1.0.4 (Released on 8/16/2022)
+## 1.0.4 (Released on <DateFormat value={new Date("Tue Aug 16 2022 11:15:14 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -97,7 +98,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
 
 * [#2272](https://github.com/marigold-ui/marigold/pull/2272) [`fb9df312`](https://github.com/marigold-ui/marigold/commit/fb9df312e50a5d4be27a528e339f0d2c5768324d) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update testing
 
-## 1.0.3 (Released on 7/28/2022)
+## 1.0.3 (Released on <DateFormat value={new Date("Thu Jul 28 2022 09:48:36 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -107,13 +108,13 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
 
 - [#2215](https://github.com/marigold-ui/marigold/pull/2215) [`247d615e`](https://github.com/marigold-ui/marigold/commit/247d615e9b127c3f2a94d0ad7561c09d11000366) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update testing
 
-## 1.0.2 (Released on 6/23/2022)
+## 1.0.2 (Released on <DateFormat value={new Date("Thu Jun 23 2022 10:09:11 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2164](https://github.com/marigold-ui/marigold/pull/2164) [`b43464fc`](https://github.com/marigold-ui/marigold/commit/b43464fce6f7e2662b27313c6f74190e8c0f540f) Thanks [@sarahgm](https://github.com/sarahgm)! - refa: update to React 18
 
-## 1.0.1 (Released on 6/3/2022)
+## 1.0.1 (Released on <DateFormat value={new Date("Fri Jun 03 2022 09:37:35 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
