+++ docs/content/changelog/config/prettier/whats-new.mdx
@@ -1,11 +1,12 @@
 ---
 title: '@marigold/prettier-config'
 caption: 'Have a look on the latest changes regarding @marigold/prettier-config'
+toc: false
 ---
 
 _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blob/main/config/prettier/CHANGELOG.md)_
 
-## 0.3.1 (Released on 3/30/2023)
+## 0.3.1 (Released on <DateFormat value={new Date("Thu Mar 30 2023 09:15:15 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
