+++ docs/content/changelog/config/tsconfig/whats-new.mdx
@@ -1,17 +1,18 @@
 ---
 title: '@marigold/tsconfig'
 caption: 'Have a look on the latest changes regarding @marigold/tsconfig'
+toc: false
 ---
 
 _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blob/main/config/tsconfig/CHANGELOG.md)_
 
-## 0.4.0 (Released on 1/9/2023)
+## 0.4.0 (Released on <DateFormat value={new Date("Mon Jan 09 2023 15:40:37 GMT+0100 (Central European Standard Time)")} dateStyle="medium" />)
 
 ### Minor Changes
 
 - [#2691](https://github.com/marigold-ui/marigold/pull/2691) [`23df35f0`](https://github.com/marigold-ui/marigold/commit/23df35f0869dfbf7cc7a921cdfd44ef1915e90fb) Thanks [@sebald](https://github.com/sebald)! - feat: explicitly set strict rules in tsconfig
 
-## 0.3.1 (Released on 12/9/2022)
+## 0.3.1 (Released on <DateFormat value={new Date("Fri Dec 09 2022 08:08:56 GMT+0100 (Central European Standard Time)")} dateStyle="medium" />)
 
 ### Patch Changes
