+++ docs/content/changelog/packages/types/whats-new.mdx
@@ -1,23 +1,24 @@
 ---
 title: '@marigold/types'
 caption: 'Have a look on the latest changes regarding @marigold/types'
+toc: false
 ---
 
 _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blob/main/packages/types/CHANGELOG.md)_
 
-## 1.1.1 (Released on 6/28/2024)
+## 1.1.1 (Released on <DateFormat value={new Date("Fri Jun 28 2024 13:44:24 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#3967](https://github.com/marigold-ui/marigold/pull/3967) [`0773aa8`](https://github.com/marigold-ui/marigold/commit/0773aa8cd6ee71faf4f0d04f80f33cbe7fc56202) Thanks [@sebald](https://github.com/sebald)! - refa: Update TypeScript and adjust `<NumericFormat>` props
 
-## 1.1.0 (Released on 10/24/2023)
+## 1.1.0 (Released on <DateFormat value={new Date("Tue Oct 24 2023 10:45:29 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Minor Changes
 
 - [#3422](https://github.com/marigold-ui/marigold/pull/3422) [`e5869b2f3`](https://github.com/marigold-ui/marigold/commit/e5869b2f3bf0f3b69a2e37f377d51786d23ccc56) Thanks [@sebald](https://github.com/sebald)! - refa: Make `<FieldBase>` work with RAC components
 
-## 1.0.1 (Released on 3/30/2023)
+## 1.0.1 (Released on <DateFormat value={new Date("Thu Mar 30 2023 09:15:15 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -25,7 +26,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
 
 - [#2891](https://github.com/marigold-ui/marigold/pull/2891) [`8972cbaca`](https://github.com/marigold-ui/marigold/commit/8972cbaca29ccce9c17a15b772ab0c7a3e921d20) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update dependency type-fest to v3.7.0
 
-## 1.0.0 (Released on 2/14/2023)
+## 1.0.0 (Released on <DateFormat value={new Date("Tue Feb 14 2023 12:38:10 GMT+0100 (Central European Standard Time)")} dateStyle="medium" />)
 
 ### Major Changes
 
@@ -41,25 +42,25 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
 
 - [#2726](https://github.com/marigold-ui/marigold/pull/2726) [`0f539b788`](https://github.com/marigold-ui/marigold/commit/0f539b788a72654e834c374810ef677c307fdadb) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update dependency @types/react to v18.0.27
 
-## 0.5.7 (Released on 12/9/2022)
+## 0.5.7 (Released on <DateFormat value={new Date("Fri Dec 09 2022 08:08:56 GMT+0100 (Central European Standard Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2674](https://github.com/marigold-ui/marigold/pull/2674) [`832da2a6`](https://github.com/marigold-ui/marigold/commit/832da2a69f9bad5adcbcc57cba3cb215dfaa51e2) Thanks [@sarahgm](https://github.com/sarahgm)! - refa: refactoring Popover with usePopover
 
-## 0.5.6 (Released on 11/24/2022)
+## 0.5.6 (Released on <DateFormat value={new Date("Thu Nov 24 2022 13:48:37 GMT+0100 (Central European Standard Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2621](https://github.com/marigold-ui/marigold/pull/2621) [`080b1fed`](https://github.com/marigold-ui/marigold/commit/080b1fedc582511e0155a9199aa3b2b995cfb609) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update dependency @types/react to v18.0.25
 
-## 0.5.5 (Released on 11/7/2022)
+## 0.5.5 (Released on <DateFormat value={new Date("Mon Nov 07 2022 13:20:21 GMT+0100 (Central European Standard Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2598](https://github.com/marigold-ui/marigold/pull/2598) [`f38ae20a`](https://github.com/marigold-ui/marigold/commit/f38ae20a23badf8cf141f582f0c20d85fdd6534b) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update dependency @types/react to v18.0.24
 
-## 0.5.4 (Released on 9/30/2022)
+## 0.5.4 (Released on <DateFormat value={new Date("Fri Sep 30 2022 15:12:18 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -67,13 +68,13 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
 
 - [#2543](https://github.com/marigold-ui/marigold/pull/2543) [`eb7e453c`](https://github.com/marigold-ui/marigold/commit/eb7e453c23f310d70d58514c24e8fe0d0cd375d4) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update dependency @types/react to v18.0.21
 
-## 0.5.3 (Released on 9/13/2022)
+## 0.5.3 (Released on <DateFormat value={new Date("Tue Sep 13 2022 09:42:53 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
 - [#2313](https://github.com/marigold-ui/marigold/pull/2313) [`75128374`](https://github.com/marigold-ui/marigold/commit/75128374e5209a24bd2b0914c1d9455c02fcfc92) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update dependency @types/react to v18.0.19
 
-## 0.5.2 (Released on 6/23/2022)
+## 0.5.2 (Released on <DateFormat value={new Date("Thu Jun 23 2022 10:09:11 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
