+++ docs/content/changelog/themes/theme-docs/whats-new.mdx
@@ -2,11 +2,12 @@
 title: '@marigold/theme-docs'
 caption: 'Have a look on the latest changes regarding @marigold/theme-docs'
 badge: new
+toc: false
 ---
 
 _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blob/main/themes/theme-docs/CHANGELOG.md)_
 
-## 2.1.1 (Released on 9/11/2024)
+## 2.1.1 (Released on <DateFormat value={new Date("Wed Sep 11 2024 14:52:49 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -15,7 +16,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
   - @marigold/system@9.0.2
   - @marigold/theme-preset@1.3.22
 
-## 2.1.0 (Released on 8/6/2024)
+## 2.1.0 (Released on <DateFormat value={new Date("Tue Aug 06 2024 13:45:06 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Minor Changes
 
@@ -28,7 +29,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
   - @marigold/system@9.0.1
   - @marigold/theme-preset@1.3.21
 
-## 2.0.5 (Released on 7/24/2024)
+## 2.0.5 (Released on <DateFormat value={new Date("Wed Jul 24 2024 11:10:39 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -41,7 +42,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
   - @marigold/system@9.0.0
   - @marigold/theme-preset@1.3.20
 
-## 2.0.4 (Released on 7/15/2024)
+## 2.0.4 (Released on <DateFormat value={new Date("Mon Jul 15 2024 09:18:40 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -50,7 +51,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
   - @marigold/system@8.0.2
   - @marigold/theme-preset@1.3.19
 
-## 2.0.3 (Released on 7/2/2024)
+## 2.0.3 (Released on <DateFormat value={new Date("Tue Jul 02 2024 13:32:49 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -59,7 +60,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
   - @marigold/system@8.0.1
   - @marigold/theme-preset@1.3.18
 
-## 2.0.2 (Released on 6/28/2024)
+## 2.0.2 (Released on <DateFormat value={new Date("Fri Jun 28 2024 13:44:24 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -68,7 +69,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
   - @marigold/system@8.0.0
   - @marigold/theme-preset@1.3.17
 
-## 2.0.1 (Released on 6/5/2024)
+## 2.0.1 (Released on <DateFormat value={new Date("Wed Jun 05 2024 14:02:30 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Patch Changes
 
@@ -79,7 +80,7 @@ _[View the full changelog on Github](https://github.com/marigold-ui/marigold/blo
   - @marigold/components@7.8.2
   - @marigold/theme-preset@1.3.16
 
-## 2.0.0 (Released on 6/5/2024)
+## 2.0.0 (Released on <DateFormat value={new Date("Wed Jun 05 2024 12:04:45 GMT+0200 (Central European Summer Time)")} dateStyle="medium" />)
 
 ### Major Changes
