+++ docs/ui/blog/PostList.tsx
@@ -1,0 +1,31 @@
+import { DateFormat, Headline, Link, Text } from '@/ui';
+import { allBlogs } from 'contentlayer/generated';
+
+export const PostList = () => {
+  // matches everything till the second line break
+  const regex = /[\s\S]*?\n[\s\S]*?\n/;
+
+  const posts = allBlogs
+    .map(post => ({
+      title: post.title,
+      date: new Date(post.date),
+      slug: post.slug,
+      introduction: post.body.raw.match(regex),
+    }))
+    .sort((a, b) => b.date.getTime() - a.date.getTime());
+
+  return (
+    <>
+      {posts.map(post => (
+        <div key={post.title} className="pb-8">
+          <Headline level={2}>
+            <Link href={`/${post.slug}`}>
+              {post.title} - <DateFormat value={post.date} dateStyle="medium" />
+            </Link>
+          </Headline>
+          <Text>{post.introduction}</Text>
+        </div>
+      ))}
+    </>
+  );
+};
